Move flatbuffer search extract methods from distributor to reuseable location
